fix running error

+++ src/main/java/com/oocl/web/parkingLot/service/impl/ParkingBoyServiceImpl.java
@@ -1,12 +1,16 @@
 package com.oocl.web.parkingLot.service.impl;
 
 import com.alibaba.fastjson.JSON;
+import com.oocl.web.parkingLot.Response.ResponseBody;
 import com.oocl.web.parkingLot.entity.ParkingBoy;
 import com.oocl.web.parkingLot.repository.ParkingBoyRepository;
 import com.oocl.web.parkingLot.service.ParkingBoyService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.trace.http.HttpTrace;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -45,7 +49,16 @@ public class ParkingBoyServiceImpl implements ParkingBoyService {
     }
 
     @Override
-    public void delete(Long id) {
-        parkingBoyRepository.deleteById(id);
+    public ResponseEntity delete(Long id) {
+        if(parkingBoyRepository.findById(id).isPresent()){
+            parkingBoyRepository.deleteById(id);
+            ParkingBoy parkingBoy = parkingBoyRepository.findById(id).get();
+
+            return ResponseEntity.status(HttpStatus.OK).body(new ResponseBody(200,"success",parkingBoy));
+        }else{
+            return ResponseEntity.status(HttpStatus.valueOf(500)).body(new ResponseBody(400,"failed",null));
+        }
+
+
     }
 }
